--- src/pages/news/News.tsx
@@ -1,5 +1,0 @@
-import React from 'react'
-
-const News = () => <div className="container">News</div>
-
-export default News
