+++ src/pages/news/index.tsx
@@ -1,1 +1,5 @@
-export { default as News } from './News'
+import React from 'react'
+
+const News = () => <div className="container">News</div>
+
+export default News
